Emails tamplates for cron tasks
At cron folder will have text files that are
used to send email messager when cron is running.
We have a cron job to run automation tasks. Names
of files are not so intuitive and yes, working
with files are not so good as working in database,
We definitely need to add an interface to edit
these files or database doing a mapper to what
event is each text related.