Template files and assets
These files are used to render helpdesk interface and
navigation to control panels. We have a lot of files
that are no more used and need to be identified and
removed. Also a mapper for identify for what each file
are used, it is desireble to create a tool for that,
allowing viewing results in place and online edit.