Email Reader
This file must NOT be exposed to the web.
email.cgi file are called by a cron job
that read email from a POP server and pipe
to it. Every piped email are converted to
a ticket based on configuration rules it can
be redirected to a specific department, or can
be blocked as spam, identify response and add
to existing ticket and so on.