Cron script
This file must NOT be exposed to the web.
Add cron.cgi to a crontab running every 5 or 10 minutes.
It is the core of system because it start everythin, from
read emails from POP, send notifications and creating services.
It is very old and MUST be refactored to be a Daemon so things
can happen instantly and must use Controller Modules to do things
with more flexibility and perfection.