Old API
This file can be exposed to the web.
I that this is not an real API but it tries
to act like one. So it MUST be refactored to a
real API with much more end points. In fact
rpc/admin.cgi can be used as a XML RPC API.
Today I will prefer to do a JSON REST API.