Build backend with Docker
Add docker files to be easier to bring all
infrastructure necessary to make Controller's
backend to work. Now with a simple command
you can get Controller work and running like
a charm.

It is the first version so it need to be improved
with some other hidden dependencies that we can
be found in future.